Fix misuse of second argument as nextState (the actual argument is prevState)
This fixes instances where, right after scrolling, a change in the
`scrollToIndex` prop will not actually scroll to the item at the passed
index correctly. This is due to the `componentDidUpdate` method
considering the previous state to actually be the next state and
looking at `scrollChangeReason` on it, seeing the old
`scrollChangeReason` caused by the user scrolling, thus not calling
`scrollTo` to actually center the element.